week-view: set the first weekday for week-header
In this commit, we set the value of first weekday for week-header.

Currently, the first weekday is being set only for the week-view, which
effects the drawing of weekdays in week header.

For fixing this, we set the value of first_weekday value of week-header
while setting the value for week-view.

https://bugzilla.gnome.org/show_bug.cgi?id=779346